requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.27
hdwallet==2.1.0
gunicorn==19.9.0
Flask-Caching==1.9.0
redis==3.5.3